Refactor package names from 'remoto' to 'remote' for consistency
